griptape_aws_bill_pdf_to_csv/requirements.txt
python-dotenv==1.0.*
griptape==1.3.0
boto3==1.35.*
pypdf==5.0.*
